[REFACT] 사용자가 작가의 프로그램을 조회하는 로직 Reservation 클래스로부터 분리
