+++ src/main/java/com/cuk/catsnap/domain/reservation/service/MemberReservationService.java
@@ -10,8 +10,6 @@ import com.cuk.catsnap.domain.photographer.service.PhotographerService;
 import com.cuk.catsnap.domain.reservation.document.ReservationTimeFormat;
 import com.cuk.catsnap.domain.reservation.dto.MonthReservationCheckListResponse;
 import com.cuk.catsnap.domain.reservation.dto.MonthReservationCheckResponse;
-import com.cuk.catsnap.domain.reservation.dto.PhotographerProgramListResponse;
-import com.cuk.catsnap.domain.reservation.dto.PhotographerProgramResponse;
 import com.cuk.catsnap.domain.reservation.dto.member.request.MemberReservationRequest;
 import com.cuk.catsnap.domain.reservation.dto.member.response.MemberReservationInformationListResponse;
 import com.cuk.catsnap.domain.reservation.dto.member.response.MemberReservationInformationResponse;
@@ -160,15 +158,6 @@ public class MemberReservationService {
 
         return PhotographerAvailableReservationTimeListResponse.from(
             photographerAvailableReservationTimeResponseArrayList);
-    }
-
-    public PhotographerProgramListResponse getPhotographerProgram(Long photographerId) {
-        List<Program> programList = programRepository.findByPhotographerIdAndDeletedFalse(
-            photographerId);
-        return PhotographerProgramListResponse.from(
-            programList.stream()
-                .map(PhotographerProgramResponse::from)
-                .toList());
     }
 
     public PhotographerReservationGuidanceResponse getPhotographerReservationGuidance(
